Tried to go the Context way for colors
