+++ src/Board.tsx
@@ -1,7 +1,8 @@
-import React, { Dispatch } from 'react';
+import React, { Dispatch, useContext } from 'react';
 import {tileIsPartOfWinningSequence} from './modules/engine';
 import {TileType, ActionType} from 'Game';
 import './Board.css';
+import { ColorsContext } from 'ColorsContext';
 
 interface Props {
   currentFrame: TileType[],
@@ -12,6 +13,7 @@ interface Props {
 
 export default function Board({currentFrame, moves, winningSequence, dispatch}: Props) {
 
+  const colors = useContext(ColorsContext);
   const rgb = (values: number[]) => {
     return `rgb(${values[0]}, ${values[1]}, ${values[2]})`;
   };
@@ -41,8 +43,7 @@ export default function Board({currentFrame, moves, winningSequence, dispatch}: 
             onClick={() => dispatch({type: 'play', payload: {tilePlayed: index}})} 
             className={getClassName(tile, index)}
             style={styleObj(tile)}>
-              {/* {tile.playedBy ? tile.playedBy : moves[index]} */}
-              {tile.playedBy ? tile.playedBy : ''}
+              {tile.playedBy ? tile.playedBy : moves[index]}
           </div>
         );
       })}
