+++ src/Game.tsx
@@ -28,29 +28,20 @@ export interface ActionType {
   }
 }
 
-const defaultColors = [
-  [180, 180, 180],
-  [16, 16, 16],
-  [240, 240, 240]
-];
+
 
 const emptyFrames = [new Array(9).fill({
     playedBy: '',
-    colorScheme: defaultColors
   })];
 
 
 function reducer(state: TileType[][], action: ActionType) {
-  console.log(state);
-  console.log(action);
   switch (action.type) {
     case 'travel':
       if (action.payload.resetTo === undefined) return emptyFrames;
-      console.log('travel');
       return state.slice(0, action.payload.resetTo + 1);
     case 'play':
-        if (action.payload.tilePlayed === undefined) return emptyFrames;
-        console.log('play');
+      if (action.payload.tilePlayed === undefined) return emptyFrames;
       return toggleTile(state, action.payload.tilePlayed);
     default: 
       return emptyFrames;
@@ -60,7 +51,6 @@ function reducer(state: TileType[][], action: ActionType) {
 const toggleTile = (frames: TileType[][], index: number) => {
   if (moveIsNotLegalAt(frames, index)) return frames;
 
-  console.log('will toggle');
   return markTileToCurrentPlayer(frames, index);
 };
 
@@ -75,8 +65,6 @@ const markTileToCurrentPlayer = (frames: TileType[][], index: number) => {
     playedBy: getCurrentTurn(frames),
     colorScheme: defaultColors
   };
-
-  // generateNewColorSchemeFor(newFrame[index]);
 
   return frames.concat([newFrame]);
 };
@@ -131,22 +119,15 @@ const lastPlayedBy = (frames: TileType[][]) => {
   return frames.length % 2 === 1 ? TurnsEnum.o : TurnsEnum.x;
 };
 
-const generateNewColorSchemeFor = async (tile: TileType) => {
-  const data = {model: 'default'};
-  const request = new Request('http://colormind.io/api/', {method: 'POST', body: JSON.stringify(data)});
-  const response = await fetch(request);
-
-  tile.colorScheme = (await response.json()).result.slice(0, 3);
-  console.log(tile.colorScheme);
-};
-
-
-
-
 export default function Game() {
 
   const [frames, dispatch] = useReducer(reducer, emptyFrames);
   const [moves, setMoves] = useState(new Array(9).fill(-Infinity));
+  const [colorSchemes, setColorSchemes] = useState(defaultColors);
+
+  useEffect(() => {
+    // fetchColorSchemes();
+  }, []);
 
   useEffect(() => {
 
@@ -159,6 +140,10 @@ export default function Game() {
     }
 
   }, [frames]);
+
+
+
+  const ColorsContext = React.createContext(colorSchemes);
 
   const playAI = (frames: TileType[][], playerMoves: number[]) => {
     const maxValue = Math.max(...playerMoves);
@@ -185,14 +170,16 @@ export default function Game() {
   }
 
   return (
-    <div className="Grid">
-      <div>
-        <h4>{gameTitle()}</h4>
-        <Board currentFrame={lastFrame(frames)} moves={moves} winningSequence={winningSequence(frames)} dispatch={dispatch} />
-        <button onClick={() => dispatch({type: 'reset', payload: {}})}>Restart game</button>
+    <ColorsContext.Provider value={colorSchemes}>
+      <div className="Grid">
+        <div>
+          <h4>{gameTitle()}</h4>
+          <Board currentFrame={lastFrame(frames)} moves={moves} winningSequence={winningSequence(frames)} dispatch={dispatch} />
+          <button onClick={() => dispatch({type: 'reset', payload: {}})}>Restart game</button>
+        </div>
+        
+        <TimeTravel frames={frames} dispatch={dispatch} />
       </div>
-      
-      <TimeTravel frames={frames} dispatch={dispatch} />
-    </div>
+    </ColorsContext.Provider>
   );
 }
